refactor: DirectionalLight component
