+++ src/components/Lights.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { useControls, folder, button } from 'leva'
 import Ground from './Ground'
-import DirectionalLight from './DirectionalLight'
+import DirectionalLight from './DirectionalLight/DirectionalLight'
 import AmbientLight from './AmbientLight'
 
 function Lights({ setBackColor }) {
